[AFF] Push 061123 1804

+++ AffiliateStoreBE/AffiliateStoreBE/Models/Product.cs
@@ -7,8 +7,8 @@ namespace AffiliateStoreBE.Models
         public Guid Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public double Cost { get; set; }
-        public double Price { get; set; }
+        public float Cost { get; set; }
+        public float Price { get; set; }
         public string Images { get; set; }
         public Guid CategoryId { get; set; }
         public int Stars { get; set; }
